requirements.txt
tqdm
pybenchmark
graphviz
torch>=0.4.0
pandas
onnx
onnx_caffe2
pretrainedmodels
torchvision
matplotlib
filelock
scikit_image
inflection
numpy
humanize
Pillow
PyQt5
scipy
skimage
tensorboardX>=1.2
tensorflow
PyYAML
pycocotools
